// config/types.ts
export interface IntegrationConfig {
  analytics:
    | { provider: "none" }
    | { provider: "google-analytics"; measurementId: string }
    | { provider: "pirsch"; code: string };
  ads:
    | { provider: "none" }
    | {
        provider: "adsterra";
        nativeScriptUrl: string;
        nativeContainerId: string;
      };
  verification: {
    google: string | null;
    bing: string | null;
  };
}